+++ check_apps_test.go
@@ -6,27 +6,27 @@ import (
 	"github.com/gomorpheus/morpheus-go-sdk"
 )
 
-func TestScriptTemplates(t *testing.T) {
+func TestCheckApps(t *testing.T) {
 	client := getTestClient(t)
 	req := &morpheus.Request{}
-	resp, err := client.ListScriptTemplates(req)
+	resp, err := client.ListCheckApps(req)
 	assertResponse(t, resp, err)
 }
 
-func TestGetScriptTemplate(t *testing.T) {
+func TestGetCheckApp(t *testing.T) {
 	client := getTestClient(t)
 	req := &morpheus.Request{}
-	resp, err := client.ListScriptTemplates(req)
+	resp, err := client.ListCheckApps(req)
 	assertResponse(t, resp, err)
 
 	// parse JSON and fetch the first one by ID
-	result := resp.Result.(*morpheus.ListScriptTemplatesResult)
+	result := resp.Result.(*morpheus.ListCheckAppsResult)
 	recordCount := result.Meta.Total
-	t.Logf("Found %d Script Templates.", recordCount)
+	t.Logf("Found %d Check Apps.", recordCount)
 	if recordCount != 0 {
 		// Get by ID
-		record := (*result.ScriptTemplates)[0]
-		resp, err = client.GetScriptTemplate(record.ID, &morpheus.Request{})
+		record := (*result.CheckApps)[0]
+		resp, err = client.GetCheckApp(record.ID, &morpheus.Request{})
 		assertResponse(t, resp, err)
 	}
 }